+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Version 1.2.1
+
+- Bug fixes
+- Improved speed by skipping waiting before cached papers
+- Improving cache lookup to prevent collecting papers missing until limit is reached
+- Allow to exclude papers
+- Add list mode
+- Add customizable log / cache file
+- Improve graph tooltip information
+- Overwork cache / log file to contain request limits
+
 ## Version 1.2.0
 
 - More granular caching
@@ -8,6 +19,7 @@
 - Skip unnecessary waiting time
 - Skip papers where citation count is zero
 - Fix maximum number of citations
+- New GUI
 
 ## Version 1.1.0